Refactoring of frontend
